requirements.txt
psutil==5.9.4
pyinstaller==5.9.0
pyusb==1.2.1
